Goodbie deadlocks =) Made it so we give tasks to cardetector with Mat, like this we no longer call functions on the same object on multiple threads that use locking mechanism. Rezults are stored in a map and only after we sync with the obj detecting  thread we use the rezults of the car detecting thread

+++ CarDetector/cvision/MovingObjectGroup.hpp
@@ -25,7 +25,7 @@ namespace cvision
 		cv::Point futurePosition_;
 		std::mutex mutexGroup_;
 		size_t nrFramesWithoutBeeingCar_;
-		size_t nrCars_;
+		uint8_t nrCars_;
 		std::atomic<size_t> cntLock = 0;
 
 		void predictNextPosition();
@@ -46,8 +46,8 @@ namespace cvision
 		MovingObjectPtr getFirstState();
 		MovingObjectPtr getLastState();
 		cv::Mat getCroppedImage(cv::Mat img);
-		void updateCarState(size_t nrCars);
-		size_t nrCars();
+		void updateCarState(uint8_t nrCars);
+		uint8_t nrCars();
 	};
 	typedef std::shared_ptr<MovingObjectGroup> MovingObjectGroupPtr;
 	typedef std::list<MovingObjectGroupPtr> MovingObjGroupPtrList;
